final part b

+++ database.py
@@ -1,29 +1,35 @@
 import pandas as pd
+import os
 
-# import game_field
-
-# print(state)
-
-df = pd.DataFrame(list())
-df.to_csv("FlagGameSave.csv")
+if not os.path.exists("FlagGameSave.csv"):
+    df = pd.DataFrame(list())
+    df.to_csv("FlagGameSave.csv")
 
 
 def create_df(number, state):
     df = pd.read_csv("FlagGameSave.csv")
     state_list = [state[i] for i in state.keys()]
-    df[number] = state_list
+    df[str(number)] = state_list
     df.to_csv("FlagGameSave.csv")
+
+
+def return_to_list(string):
+    list = []
+    for i in range(len(string)):
+        if string[i] == "]":
+            list.append("".join(string[1:i]))
+    return list
+
 
 
 def retrieve_state(number, state):
     df = pd.read_csv('FlagGameSave.csv')
-    print(df)
-    df.to_dict()
-    # print(df)
-    df = df[number]
 
+    new_state = list(df[str(number)])
     count = 0
     for i in state.keys():
-        state[i] = df[count]
+        state[i] = eval(new_state[count])
         count += 1
+    print(state)
+
     return state
